add missing tests for RequestMessage
